requirements.txt
flask
vanna[postgres,mysql,openai,pinecone]
db-dtypes
python-dotenv
pinecone
fastembed
openai
langsmith
